refactor(sftp): extract shared transfer-feedback helpers
Pull runTransfer, the Transfer-Queue seed, and the path/error helpers out of
useFileSystem into a shared transferFeedback module so the SFTP and session
file-browser hooks drive an identical transfer contract (pending toast,
single-terminal-toast rule, seed row) over their respective dedicated-channel
commands. Behaviour-preserving; prep for the SFTP browser convergence (#2421).
