+++ src/hooks/useFileSystem.ts
@@ -9,71 +9,9 @@ import {
   sftpRename,
   sftpWriteFileContent,
   vscodeOpenRemote,
-  TransferTerminalError,
 } from "@/services/api";
 import { FileEntry } from "@/types/connection";
-import { toast } from "@/components/ui";
-import { frontendLog } from "@/utils/frontendLog";
-import { dispatchTransferIntentBestEffort } from "@/store/transfersBridge";
-
-/** Basename of a POSIX/Windows path (the file name the queue row displays). */
-function baseName(path: string): string {
-  const parts = path.replace(/\\/g, "/").split("/");
-  return parts[parts.length - 1] || path;
-}
-
-/** Extract a human-readable message from an unknown transfer error. */
-function transferErrorMessage(error: unknown): string {
-  if (error instanceof Error) return error.message;
-  if (typeof error === "string") return error;
-  return String(error);
-}
-
-/**
- * Run an SFTP transfer with user feedback: a pending toast shown while the
- * transfer runs (audit gap D2 — transfers must never fail silently).
- *
- * The **terminal** success/error toast is owned exclusively by the
- * `transfer-progress` event path (`useTransferEvents`, #1286) so a single
- * transfer produces exactly one terminal toast. This helper therefore does not
- * emit its own success/error toast for a transfer that reached the backend:
- *   - success → dismiss the pending toast (the event path raises the "Downloaded
- *     …/Uploaded …" success toast);
- *   - a {@link TransferTerminalError} (`done`/`cancelled`/`error` from the
- *     transfer channel) → dismiss the pending toast; the event path already
- *     surfaced the outcome (and stays quiet on cancel).
- *
- * An **early** failure that never produced a transfer event (invalid session,
- * permission error thrown before the transfer starts, an sftp→sftp temp-file
- * paste leg that fails synchronously) is still surfaced here via `toast.error`,
- * since no `transfer-progress` event will cover it. The rejection is always
- * swallowed so callers never produce an unhandled rejection. Returns whether
- * the transfer succeeded.
- */
-async function runTransfer(
-  label: string,
-  action: () => Promise<unknown>,
-  messages: { loading: string }
-): Promise<boolean> {
-  const toastId = toast.loading(messages.loading);
-  try {
-    await action();
-    // Success toast is raised by the transfer-progress event path (#1286).
-    toast.dismiss(toastId);
-    return true;
-  } catch (error) {
-    if (error instanceof TransferTerminalError) {
-      // The event path already surfaced this (success/error toast) or is
-      // intentionally quiet (cancel) — just clear the pending toast.
-      toast.dismiss(toastId);
-      return false;
-    }
-    const message = transferErrorMessage(error);
-    frontendLog("sftp_transfer", `${label} failed: ${message}`);
-    toast.error(`${label} failed: ${message}`, { id: toastId });
-    return false;
-  }
-}
+import { runTransfer, seedTransferQueueRow } from "./transferFeedback";
 
 /**
  * Hook for SFTP file system operations.
@@ -90,38 +28,18 @@ export function useFileSystem() {
   // Wrap the transfer helpers so every transfer seeds a `queued` Transfer Queue
   // row from the id the start command returns — the panel then opens without
   // waiting for a `transfer-progress` event, which can be dropped/delayed under
-  // memory pressure (#1632). The row is keyed by the backend `transferId`, so a
-  // later event upserts (never duplicates) it. Since #2229 the `transfers` region
-  // is authoritative, so the seed is a reliable client `transfer.seed` intent
-  // against the shared store (idempotent server-side — it never overwrites an
-  // already-advanced row); a bridge hiccup is swallowed and logged.
+  // memory pressure (#1632). See {@link seedTransferQueueRow}.
   const startDownload = useCallback(
     (sessionId: string, remotePath: string, localPath: string) =>
       sftpDownload(sessionId, remotePath, localPath, (transferId) =>
-        dispatchTransferIntentBestEffort("transfer.seed", {
-          seed: {
-            id: transferId,
-            sessionId,
-            direction: "download",
-            name: baseName(remotePath),
-            path: remotePath,
-          },
-        })
+        seedTransferQueueRow({ transferId, sessionId, direction: "download", remotePath })
       ),
     []
   );
   const startUpload = useCallback(
     (sessionId: string, localPath: string, remotePath: string) =>
       sftpUpload(sessionId, localPath, remotePath, (transferId) =>
-        dispatchTransferIntentBestEffort("transfer.seed", {
-          seed: {
-            id: transferId,
-            sessionId,
-            direction: "upload",
-            name: baseName(remotePath),
-            path: remotePath,
-          },
-        })
+        seedTransferQueueRow({ transferId, sessionId, direction: "upload", remotePath })
       ),
     []
   );
